reafactor: 반응결과 컴포넌트 분리
